test: keep CLI cold-start coverage within package limits
